+++ api/src/routes/index.js
@@ -19,9 +19,10 @@ router.get('/extents/*', (req, res) => {
   const url = `${OSMESA_API}/extents/${req.params[0]}`
   req.pipe(request(url)).pipe(res)
 })
+router.get('/users', users.list)
+router.get('/users/stats', users.stats)
 router.get('/users/:id', user.get)
 router.put('/users/:id', user.put)
-router.get('/users', users)
 router.get('/campaigns/:id', campaign)
 router.get('/campaigns', campaigns)
 router.get('/topstats', topstats)